Move argument parsing to a separate file
